requirements.txt
torch>=1.13
numpy
matplotlib
pyyaml
tqdm
